Split note transaction tests into separate files (importing shared tests)
