+++ apps/e2e/integration/note-transactions/shared_tests.ts
@@ -1,9 +1,10 @@
+import { type Page, test } from "@playwright/test";
+
 import { NoteState, NoteTempState } from "@librocco/shared";
-import { Page, test } from "@playwright/test";
 
-import { baseURL } from "../constants";
+import { baseURL } from "../../constants";
 
-import { getDashboard } from "../helpers";
+import { getDashboard } from "../../helpers";
 
 const book1 = {
 	isbn: "1234567890",
@@ -26,18 +27,6 @@ const book2 = {
 	editedBy: "No one",
 	outOfPrint: false
 };
-
-const book3 = {
-	isbn: "1234567892",
-	title: "Big Bad Wolf",
-	price: 10,
-	year: "1999",
-	authors: "Walter White",
-	publisher: "Penguin Random House",
-	editedBy: "Saul Goodman",
-	outOfPrint: false
-};
-
 const createNoteForView = (page: Page, view: "inbound" | "outbound") => {
 	const sidebar = getDashboard(page).sidebar();
 
@@ -51,27 +40,8 @@ const createNoteForView = (page: Page, view: "inbound" | "outbound") => {
 	return sidebar.linkGroup("New Warehouse").createNote();
 };
 
-const runNoteTransactionTests = (view: "inbound" | "outbound") => {
-	test.beforeEach(async ({ page }) => {
-		// Load the app
-		await page.goto(baseURL);
-
-		const dashboard = getDashboard(page);
-		const sidebar = dashboard.sidebar();
-
-		// Wait for the app to become responsive (when the default view is loaded)
-		await dashboard.waitFor();
-
-		// Create a warehouse first (to which we can add the notes, in case of inbound)
-		await sidebar.createWarehouse();
-
-		// Navigate to the correct view
-		await dashboard.navigate(view);
-
-		// Create a new note to work with
-		await createNoteForView(page, view);
-	});
-
+export const runCommonTransactionTests = (view: "inbound" | "outbound") => {
+	/** @TODO should be tested view-specific */
 	test("should add a transaction to the note by filling out the book form on 'Create' button", async ({ page }) => {
 		const dashboard = getDashboard(page);
 
@@ -141,18 +111,16 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 		]);
 	});
 
-	test("should aggregate the quantity of the same book", async ({ page }) => {
+	test("should aggregate the quantity for the same isbn", async ({ page }) => {
 		const dashboard = getDashboard(page);
 
 		const content = dashboard.content();
 		const scanField = content.scanField();
 		const entries = content.entries(view);
 
-		// Create book 1 transaction (quanitity 1)
-		await scanField.add(book1.isbn);
-
-		// Create book 2 transaction (quanitity 1)
-		await scanField.add(book2.isbn);
+		// Create two transactions with different isbns (quanitity 1 each)
+		await scanField.add("1234567890");
+		await scanField.add("1234567891");
 
 		// Check that both books are in the entries table
 		// (by not using 'strict: true', we're asserting only by values we care about)
@@ -167,10 +135,10 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 			}
 		]);
 
-		// Create another transaction for book 1
-		await scanField.add(book1.isbn);
+		// Add another transaction for "1234567890"
+		await scanField.add("1234567890");
 
-		// No new transaction should be added, but the quantity of book 1 should be increased
+		// No new transaction should be added, but the quantity of "1234567890" should be increased
 		await entries.assertRows([
 			{
 				isbn: "1234567890",
@@ -192,13 +160,15 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 		const scanField = content.scanField();
 		const entries = content.entries(view);
 
+		/***** @TODO this part should be replaced by programmatically creating a book data entry */
 		// Add transaction with book 1 (filling the form with book data)
 		await scanField.create();
 		await bookForm.fillBookData(book1);
 		await bookForm.submit("click");
 
-		// Add another note
+		// Add another note (when programmatic setup is supported, we will be testing with the default note )
 		await createNoteForView(page, view);
+		/***** @TODO this part should be replaced by programmatically creating a book data entry */
 
 		// Add book 1 again (this time using only isbn and 'Add' button)
 		await scanField.add(book1.isbn);
@@ -259,11 +229,11 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 		const entries = content.entries(view);
 
 		// We're adding books in non-aplphabetical order to check if they're sorted correctly
-		const books = [book2, book3, book1];
+		const isbns = ["1234567891", "1234567890", "1234567892"];
 
 		// Add transactions
-		for (const book of books) {
-			await scanField.add(book.isbn);
+		for (const isbn of isbns) {
+			await scanField.add(isbn);
 		}
 
 		// Check that the transactions are sorted by isbn
@@ -277,10 +247,10 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 		const scanField = content.scanField();
 		const entries = content.entries(view);
 
-		// Add three books
-		const books = [book1, book2, book3];
-		for (const book of books) {
-			await scanField.add(book.isbn);
+		// Add three transactions
+		const isbns = ["1234567890", "1234567891", "1234567892"];
+		for (const isbn of isbns) {
+			await scanField.add(isbn);
 		}
 
 		// Delete the second transaction
@@ -307,10 +277,10 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 		const scanField = content.scanField();
 		const entries = content.entries(view);
 
-		// Add three books
-		const books = [book1, book2, book3];
-		for (const book of books) {
-			await scanField.add(book.isbn);
+		// Add three transactions
+		const isbns = ["1234567890", "1234567891", "1234567892"];
+		for (const isbn of isbns) {
+			await scanField.add(isbn);
 		}
 
 		// Select all transactions
@@ -337,10 +307,10 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 		const scanField = content.scanField();
 		const entries = content.entries(view);
 
-		// Add three books
-		const books = [book1, book2, book3];
-		for (const book of books) {
-			await scanField.add(book.isbn);
+		// Add three transactions
+		const isbns = ["1234567890", "1234567891", "1234567892"];
+		for (const isbn of isbns) {
+			await scanField.add(isbn);
 		}
 
 		// Select all transactions
@@ -372,16 +342,12 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 	test.skip("should not allow committing a note with 0-quantity transaction(s)", async ({ page }) => {
 		const dashboard = getDashboard(page);
 
-		const bookForm = dashboard.bookForm();
-
 		const content = dashboard.content();
 		const entries = content.entries(view);
 		const statePicker = content.statePicker();
 
 		// Add a transaction with 0 quantity
-		await content.scanField().create();
-		await bookForm.fillBookData(book1);
-		await bookForm.submit("click");
+		await content.scanField().add("1234567890");
 		await entries.row(0).setQuantity(0);
 
 		// Try and commit the note
@@ -392,11 +358,3 @@ const runNoteTransactionTests = (view: "inbound" | "outbound") => {
 		await statePicker.assertState(NoteTempState.Committing);
 	});
 };
-
-test.describe("Inbound note view", () => {
-	runNoteTransactionTests("inbound");
-});
-
-test.describe("Outbound note view", () => {
-	runNoteTransactionTests("outbound");
-});
